requirements.txt
openai>=1.12.0
pydantic>=2.6.0
streamlit>=1.30.0
python-dotenv>=1.0.0